refactor: move config pkg out internal, required in higher level pkgs
