+++ cmd/gitops/beta/cmd.go
@@ -4,7 +4,7 @@ import (
 	"github.com/go-resty/resty/v2"
 	"github.com/spf13/cobra"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/beta/run"
-	"github.com/weaveworks/weave-gitops/cmd/internal/config"
+	"github.com/weaveworks/weave-gitops/cmd/gitops/config"
 )
 
 func GetCommand(opts *config.Options, client *resty.Client) *cobra.Command {
