+++ cmd/gitops/delete/cmd.go
@@ -3,8 +3,8 @@ package delete
 import (
 	"github.com/go-resty/resty/v2"
 	"github.com/spf13/cobra"
+	"github.com/weaveworks/weave-gitops/cmd/gitops/config"
 	"github.com/weaveworks/weave-gitops/cmd/gitops/delete/clusters"
-	"github.com/weaveworks/weave-gitops/cmd/internal/config"
 )
 
 func DeleteCommand(opts *config.Options, client *resty.Client) *cobra.Command {
