+++ cmd/gitops/get/templates/terraform/cmd.go
@@ -9,7 +9,7 @@ import (
 	"k8s.io/cli-runtime/pkg/printers"
 
 	"github.com/weaveworks/weave-gitops/cmd/gitops/cmderrors"
-	"github.com/weaveworks/weave-gitops/cmd/internal/config"
+	"github.com/weaveworks/weave-gitops/cmd/gitops/config"
 	"github.com/weaveworks/weave-gitops/pkg/adapters"
 	"github.com/weaveworks/weave-gitops/pkg/templates"
 )
